+++ examples/zed_oc_control_example.cpp
@@ -48,22 +48,22 @@ CamControl activeControl = Brightness;
 
 // ----> Global functions to control settings
 // Rescale the images according to the selected resolution to better display them on screen
-void showImage( std::string name, cv::Mat& img, sl_oc::RESOLUTION res );
+void showImage( std::string name, cv::Mat& img, sl_oc::video::RESOLUTION res );
 
 // Handle Keyboard
-void handleKeyboard( sl_oc::VideoCapture &cap, int key );
+void handleKeyboard( sl_oc::video::VideoCapture &cap, int key );
 
 // Change active control
 void setActiveControl( CamControl control );
 
 // Set new value for the active control
-void setControlValue( sl_oc::VideoCapture &cap, int value );
+void setControlValue( sl_oc::video::VideoCapture &cap, int value );
 
 // '+' or '-' pressed
-void changeControlValue( sl_oc::VideoCapture &cap, bool increase );
+void changeControlValue( sl_oc::video::VideoCapture &cap, bool increase );
 
 // 'a' or 'A' pressed to enable automatic WhiteBalanse or Gain/Exposure
-void toggleAutomaticControl( sl_oc::VideoCapture &cap );
+void toggleAutomaticControl( sl_oc::video::VideoCapture &cap );
 // <---- Global functions to control settings
 
 // The main function
@@ -71,15 +71,15 @@ int main(int argc, char *argv[])
 {
     sl_oc::VERBOSITY verbose = sl_oc::VERBOSITY::INFO;
 
-    // ----> 1) Set Video parameters
-    sl_oc::VideoParams params;
-    params.res = sl_oc::RESOLUTION::HD2K;
-    params.fps = sl_oc::FPS::FPS_15;
+    // ----> Set Video parameters
+    sl_oc::video::VideoParams params;
+    params.res = sl_oc::video::RESOLUTION::HD2K;
+    params.fps = sl_oc::video::FPS::FPS_15;
     params.verbose = verbose;
     // <---- Set Video parameters
 
-    // ----> 2) Create Video Capture
-    sl_oc::VideoCapture cap(params);
+    // ----> Create Video Capture
+    sl_oc::video::VideoCapture cap(params);
     if( !cap.initializeVideo(-1) )
     {
         std::cerr << "Cannot open camera video capture" << std::endl;
@@ -93,30 +93,33 @@ int main(int argc, char *argv[])
     // Set the default camera control setting
     setActiveControl( Brightness );
 
+    uint64_t last_ts=0;
+
     // Infinite video grabbing loop
     while (1)
     {
         // 3) Get last available frame
-        const sl_oc::Frame* frame = cap.getLastFrame();
+        const sl_oc::video::Frame frame = cap.getLastFrame();
 
-        // ----> 4) If the frame is valid we can display it
-        if(frame != nullptr)
+        // ----> If the frame is valid we can display it
+        if(frame.data!=nullptr && frame.timestamp!=last_ts)
         {
 #if 0
-            // ----> 4.a) Video Debug information
-            static uint64_t last_ts=0;
-            std::cout << std::setprecision(9) << "[" << frame->frame_id << "] Ts: " <<  static_cast<double>(frame->timestamp)/1e9 << " sec" << std::endl;
+            // ----> Video Debug information
+
+            std::cout << std::setprecision(9) << "[" << frame.frame_id << "] Ts: " <<  static_cast<double>(frame.timestamp)/1e9 << " sec" << std::endl;
             if( last_ts!=0 )
             {
-                double dt = (frame->timestamp - last_ts)/1e9;
+                double dt = (frame.timestamp - last_ts)/1e9;
                 std::cout << std::setprecision(9) << " * dT: " << dt << " sec - FPS: " << 1./dt <<  std::endl;
             }
-            last_ts = frame->timestamp;
+
             // <---- Video Debug information
 #endif
+            last_ts = frame.timestamp;
 
-            // ----> 4.b)Conversion from YUV 4:2:2 to BGR for visualization
-            cv::Mat frameYUV = cv::Mat( frame->height, frame->width, CV_8UC2, frame->data );
+            // ----> Conversion from YUV 4:2:2 to BGR for visualization
+            cv::Mat frameYUV = cv::Mat( frame.height, frame.width, CV_8UC2, frame.data );
             cv::Mat frameBGR;
             cv::cvtColor(frameYUV,frameBGR,cv::COLOR_YUV2BGR_YUYV);
             // <---- Conversion from YUV 4:2:2 to BGR for visualization
@@ -126,7 +129,7 @@ int main(int argc, char *argv[])
         }
         // <---- If the frame is valid we can display it
 
-        // ----> 5) Keyboard handling
+        // ----> Keyboard handling
         int key = cv::waitKey( 5 );
 
         if( key != -1 )
@@ -143,7 +146,7 @@ int main(int argc, char *argv[])
 }
 
 // Handle Keyboard
-void handleKeyboard( sl_oc::VideoCapture &cap, int key )
+void handleKeyboard( sl_oc::video::VideoCapture &cap, int key )
 {
     if(key >= '0' && key <= '9')
     {
@@ -291,7 +294,7 @@ void setActiveControl( CamControl control )
 }
 
 // Set new value for the active control
-void setControlValue(sl_oc::VideoCapture &cap, int value )
+void setControlValue(sl_oc::video::VideoCapture &cap, int value )
 {
     int newValue;
     switch( activeControl )
@@ -356,7 +359,7 @@ void setControlValue(sl_oc::VideoCapture &cap, int value )
 }
 
 // '+' or '-' pressed
-void changeControlValue( sl_oc::VideoCapture &cap, bool increase )
+void changeControlValue( sl_oc::video::VideoCapture &cap, bool increase )
 {
     int curValue=0;
     switch( activeControl )
@@ -381,43 +384,43 @@ void changeControlValue( sl_oc::VideoCapture &cap, bool increase )
 
     case Gain:
     {
-        int curValueLeft = cap.getGain(sl_oc::CAM_SENS_POS::LEFT);
-        int curValueRight = cap.getGain(sl_oc::CAM_SENS_POS::RIGHT);
+        int curValueLeft = cap.getGain(sl_oc::video::CAM_SENS_POS::LEFT);
+        int curValueRight = cap.getGain(sl_oc::video::CAM_SENS_POS::RIGHT);
 
         if(increase)
         {
-            cap.setGain(sl_oc::CAM_SENS_POS::LEFT,++curValueLeft);
-            cap.setGain(sl_oc::CAM_SENS_POS::RIGHT,++curValueRight);
+            cap.setGain(sl_oc::video::CAM_SENS_POS::LEFT,++curValueLeft);
+            cap.setGain(sl_oc::video::CAM_SENS_POS::RIGHT,++curValueRight);
         }
         else
         {
-            cap.setGain(sl_oc::CAM_SENS_POS::LEFT,--curValueLeft);
-            cap.setGain(sl_oc::CAM_SENS_POS::RIGHT,--curValueRight);;
+            cap.setGain(sl_oc::video::CAM_SENS_POS::LEFT,--curValueLeft);
+            cap.setGain(sl_oc::video::CAM_SENS_POS::RIGHT,--curValueRight);;
         }
 
-        std::cout << "New Left Gain value: " << cap.getGain(sl_oc::CAM_SENS_POS::LEFT) << std::endl;
-        std::cout << "New Right Gain value: " << cap.getGain(sl_oc::CAM_SENS_POS::RIGHT) << std::endl;
+        std::cout << "New Left Gain value: " << cap.getGain(sl_oc::video::CAM_SENS_POS::LEFT) << std::endl;
+        std::cout << "New Right Gain value: " << cap.getGain(sl_oc::video::CAM_SENS_POS::RIGHT) << std::endl;
     }
         break;
 
     case Exposure:
     {
-        int curValueLeft = cap.getExposure(sl_oc::CAM_SENS_POS::LEFT);
-        int curValueRight = cap.getExposure(sl_oc::CAM_SENS_POS::RIGHT);
+        int curValueLeft = cap.getExposure(sl_oc::video::CAM_SENS_POS::LEFT);
+        int curValueRight = cap.getExposure(sl_oc::video::CAM_SENS_POS::RIGHT);
 
         if(increase)
         {
-            cap.setExposure(sl_oc::CAM_SENS_POS::LEFT,++curValueLeft);
-            cap.setExposure(sl_oc::CAM_SENS_POS::RIGHT,++curValueRight);
+            cap.setExposure(sl_oc::video::CAM_SENS_POS::LEFT,++curValueLeft);
+            cap.setExposure(sl_oc::video::CAM_SENS_POS::RIGHT,++curValueRight);
         }
         else
         {
-            cap.setExposure(sl_oc::CAM_SENS_POS::LEFT,--curValueLeft);
-            cap.setExposure(sl_oc::CAM_SENS_POS::RIGHT,--curValueRight);;
+            cap.setExposure(sl_oc::video::CAM_SENS_POS::LEFT,--curValueLeft);
+            cap.setExposure(sl_oc::video::CAM_SENS_POS::RIGHT,--curValueRight);;
         }
 
-        std::cout << "New Left Exposure value: " << cap.getExposure(sl_oc::CAM_SENS_POS::LEFT) << std::endl;
-        std::cout << "New Right Exposure value: " << cap.getExposure(sl_oc::CAM_SENS_POS::RIGHT) << std::endl;
+        std::cout << "New Left Exposure value: " << cap.getExposure(sl_oc::video::CAM_SENS_POS::LEFT) << std::endl;
+        std::cout << "New Right Exposure value: " << cap.getExposure(sl_oc::video::CAM_SENS_POS::RIGHT) << std::endl;
     }
         break;
 
@@ -457,7 +460,7 @@ void changeControlValue( sl_oc::VideoCapture &cap, bool increase )
 }
 
 // 'a' or 'A' pressed to enable automatic WhiteBalanse or Gain/Exposure
-void toggleAutomaticControl( sl_oc::VideoCapture &cap )
+void toggleAutomaticControl( sl_oc::video::VideoCapture &cap )
 {
     if(activeControl == WhiteBalance)
     {
@@ -477,21 +480,21 @@ void toggleAutomaticControl( sl_oc::VideoCapture &cap )
 }
 
 // Rescale the images according to the selected resolution to better display them on screen
-void showImage( std::string name, cv::Mat& img, sl_oc::RESOLUTION res )
+void showImage( std::string name, cv::Mat& img, sl_oc::video::RESOLUTION res )
 {
     cv::Mat resized;
     switch(res)
     {
     default:
-    case sl_oc::RESOLUTION::VGA:
+    case sl_oc::video::RESOLUTION::VGA:
         resized = img;
         break;
-    case sl_oc::RESOLUTION::HD720:
+    case sl_oc::video::RESOLUTION::HD720:
         name += " [Resize factor 0.6]";
         cv::resize( img, resized, cv::Size(), 0.6, 0.6 );
         break;
-    case sl_oc::RESOLUTION::HD1080:
-    case sl_oc::RESOLUTION::HD2K:
+    case sl_oc::video::RESOLUTION::HD1080:
+    case sl_oc::video::RESOLUTION::HD2K:
         name += " [Resize factor 0.4]";
         cv::resize( img, resized, cv::Size(), 0.4, 0.4 );
         break;