+++ changelog.md
@@ -1,12 +1,15 @@
 # Changelog
 
-v0.1.1 - 2012 06 10
+v0.2 - 2012 06 10
 -------------------
+* Fix issue downloading camera settings for the rectification example
 * Documentation refactoring
 * New "sl_oc::video" namespace
 * New "sl_oc::sensors" namespace
 * New "sl_oc::sensors::data" namespace
 * New "sl_oc::sensors::usb" namespace
+* Sensors data and image data are now returned as reference instead of pointer
+* Improved sensors data validity field
 
 v0.1 - 2012 06 04
 -----------------